Reconfigure PuTTY colour scheme from Unix host.
This is possibly related to: /wishlist/colour-scheme.html

These colour schemes are not stored in the PuTTY session file but are
instead "dynamically" configured from the host.

This script can send the required control sequences suitable for any
modern version of PuTTY or the Linux console.

It can calculate several standard forms and contains a default PuTTY
and XTerm mappings. In addition it can extract the colour scheme from
a Windows PuTTY registry dump, a Unix session file and a few other
file types.