Clean up VT52 leftovers from the 256 colours patch
Some references to the unused flag vt52_bold are removed.

Remove the messing with the 'blink_is_real' flag as it's no longer
needed for the Atari colours now the flag no longer overlaps with the
bright background.

Add the MiNT blink codes now the the flag will work.

Also, the wrap mode should be set to "2" for Atari compatibility.